cleanup temporary/useless things, output rendering changes

+++ service/lhv-tools/postinst/postinstall.sh
@@ -1,19 +1,16 @@
 #!/bin/sh
 
+export TERM="vt100"
 . /etc/include/choupi
 wwwroot="var/www"
 toolsdir="../tmp/lhv-tools"
 
 
 echo "${INFO} \"LHV tools\" prerequisites installation ..."
-if [ -d ${toolsdir} ]; then
-	rm -fr ${toolsdir}
-fi
+if [ -d ${toolsdir} ]; then rm -fr ${toolsdir}; fi
 mkdir ${toolsdir}
 
-#link="https://lehollandaisvolant.net/tout/tools/tools.tar.7z"
-#link="http://192.168.1.87:40080/tools.tar.7z"
-link="http://192.168.1.17:8000/tools.tar.7z"
+link="https://lehollandaisvolant.net/tout/tools/tools.tar.7z"
 ${FETCH} -o ${toolsdir}/$(basename ${link}) ${link} 
 if [ $? -ne 0 ]; then
 	echo -e "${ERROR} \"LHV tools\" download failed.\nExit."
@@ -28,23 +25,33 @@ fi
 # to avoid space disc consumption, download and decompression are made in the
 # tmp/ directory of smolBSD, on the host file system.
 
-echo -n "${ARROW} un-7zipping ... "
+echo -n "${ARROW} un-7zipping |"
 # Unlike the "tar" command, "7z" is not necessarily installed everywhere. So it's
 # installed on the microvm with "ADDPKGS" in options.mk and used here.
-usr/pkg/bin/7z e -o${toolsdir} ${toolsdir}/$(basename ${link}) >dev/null 2>&1
+usr/pkg/bin/7z e -o${toolsdir} ${toolsdir}/$(basename ${link}) 2>&1 | awk '{printf "*"; fflush()}'
 if [ $? -eq 0 ]; then
-	echo "done"
+	echo "| done"
 else
-	echo "${ERROR} failed.\nExit"
+	echo -e "| ${ERROR} failed.\nExit"
 	. etc/include/shutdown
 fi
 
-echo -n "${ARROW} un-taring ... "
-tar -xvf ${toolsdir}/$(basename ${link%.7z}) -C ${wwwroot} >dev/null 2>&1
+echo -n "${ARROW} un-taring |"
+tar -xvf ${toolsdir}/$(basename ${link%.7z}) -C ${wwwroot} 2>&1 | awk '{printf "*"; fflush()}'
 if [ $? -eq 0 ]; then
-	echo "done"
+	echo "| done"
 else
-	echo "${ERROR} failed.\nExit"
+	echo -e "| ${ERROR} failed.\nExit"
+	. etc/include/shutdown
+fi
+
+echo -n "${ARROW} fix for icon in the header "
+find ${wwwroot} -name index.php -exec sed -i 's#/index/logo-no-border.png#../0common/lhv-384x384.png#g' {} \; 2>&1 | awk '{printf "*"; fflush()}'
+if [ $? -eq 0 ]; then
+	echo ""
+	break
+else
+	echo -e "${ERROR} failed.\nExit"
 	. etc/include/shutdown
 fi
